GameTime, architecture overhaul, DPS
+[ADD] Added GameTime, IUpdateable and IRenderable now provide this
object
+[MOD] Complete overhaul of all update (tick) and render methods
+[MOD] Damage now is measured in DPS (damage per second), AttackBehavior
now implements it this way
+[MOD] Fixed TeleportMoveBehavior#follow, also improved its structure

+++ src/de/cormag/projectf/entities/properties/offensive/ICanAttack.java
@@ -8,9 +8,9 @@ package de.cormag.projectf.entities.properties.offensive;
  */
 public interface ICanAttack extends ICanOffensive {
 	/**
-	 * Gets the attack power of this object.
+	 * Gets the attack power of this object as DPS (damage per second).
 	 * 
-	 * @return The attack power of this object
+	 * @return The attack power of this object as DPS (damage per second)
 	 */
-	public int getAttackPower();
+	public float getAttackPower();
 }
